+++ sparkle/CLI/cli.py
@@ -37,10 +37,16 @@ def main() -> None:
         command_file = module_path / f"{command}.py"
         if command in possible_commands:
             break
+
     if command_file.is_file():
         os.system(f"python3 {command_file} {' '.join(args)}")
+    elif command_file.with_suffix(".sh").is_file():
+        script_path = command_file.with_suffix(".sh")
+        script_path.chmod(0o755)  # Ensure execution rights with shipment
+        os.system(f"source {script_path}")
+        print("Sparkle autocomplete activated!")
     else:
-        print(f"Does not understand command {command}")
+        print(f"Sparkle does not understand command <{command}>")
 
 
 if __name__ == "__main__":